requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==3.1.0
idna==2.10
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.4.2
mercurial==5.6.1
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.1.0
protobuf==3.14.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==4.19.25
PySide2==5.15.2
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
shiboken2==5.15.2
six==1.15.0
TBB==0.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
torch==1.7.1
torchvision==0.8.2
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
youtube-dl==2021.4.7
youtube-dlc==2020.11.11.post3
torchlars==0.1.2